requirements.txt
streamlit
pandas
plotly
numpy
scikit-learn
transformers
sentence-transformers
openai
python-dotenv
pydantic
cohere
lightgbm
fastapi
uvicorn
requests
aiohttp

